+++ response/vectorQuery.go
@@ -7,8 +7,8 @@ import (
 
 type QueryMatch struct {
 	ID           *string                `json:"id,omitempty"`
-	Score        *float32               `json:"score,omitempty"`
-	Values       []float32              `json:"values,omitempty"`
+	Score        *float64               `json:"score,omitempty"`
+	Values       []float64              `json:"values,omitempty"`
 	SparseValues *VectorSparseValues    `json:"sparseValues,omitempty"`
 	Metadata     map[string]interface{} `json:"metadata,omitempty"`
 }
